requirements.txt
aiosqlite
discord.py
requests-cache
typing
helpers